Add node auth requirements (#5)
apiserver's Node authorization mode requires specific organization and
common names be set in the certificate in order to authorize a "node".
This adds `O=system:nodes` to allow the application wielding this
certificate to be a kubernetes node. It also adds the
`CN=system.node.<hostname>` which is a documented requirement.